Split SampleHarmonicProblems
Rewrite the implementation of the PlaneWave test case to use PlaneWave
