+++ applications/DG-Max/CMakeLists.txt
@@ -19,6 +19,7 @@ if(hpGEM_USE_COMPLEX_PETSC AND hpGEM_USE_SLEPC)
 		Algorithms/DivDGMaxEigenvalue.cpp
 
 		ProblemTypes/Harmonic/SampleHarmonicProblems.cpp
+		ProblemTypes/Harmonic/PlaneWaveReflectionProblem.cpp
 		ProblemTypes/Time/DummyTestProblem.cpp
 		ProblemTypes/Time/SampleTestProblems.cpp
 		Utils/BandstructureGNUPlot.cpp
