+++ applications/DG-Max/ProblemTypes/Harmonic/SampleHarmonicProblems.cpp
@@ -7,7 +7,7 @@ This code is distributed using BSD 3-Clause License. A copy of which can found
 below.
 
 
-Copyright (c) 2018, Univesity of Twenete
+Copyright (c) 2018, University of Twente
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without modification,
@@ -71,82 +71,4 @@ LinearAlgebra::SmallVectorC<3> SarmanyHarmonicProblem::sourceTerm(
 template class ConstantHarmonicProblem<2>;
 template class ConstantHarmonicProblem<3>;
 
-template <std::size_t dim>
-LinearAlgebra::SmallVectorC<dim> PlaneWaveReflectionProblem<dim>::exactSolution(
-    const Geometry::PointPhysical<dim> &point) const {
-
-    using namespace std::complex_literals;
-    LinearAlgebra::SmallVectorC<dim> result;
-    result[1] = 1.0;
-
-    double x = point.getCoordinate(0);
-
-    if (point.getCoordinates()[0] < interfacePosition_) {
-        result *= (incidentPhasor_ * std::exp(1i * k1_ * x) +
-                   reflectionPhasor_ * std::exp(-1i * k1_ * x));
-    } else {
-        result *= transmissionPhasor_ * std::exp(1i * k2_ * x);
-    }
-    return result;
-}
-
-template <std::size_t dim>
-LinearAlgebra::SmallVectorC<dim>
-    PlaneWaveReflectionProblem<dim>::exactSolutionCurl(
-        const Geometry::PointPhysical<dim> &point) const {
-
-    double x = point.getCoordinate(0);
-
-    using namespace std::complex_literals;
-    LinearAlgebra::SmallVectorC<dim> result;
-    LinearAlgebra::SmallVectorC<dim> kdir;
-    kdir[0] = 1.0;
-    result[1] = 1.0;
-    // Curl E0 exp(ikx) = ik x E0 exp(ikx)
-    result = kdir.crossProduct(result);
-
-    if (x < interfacePosition_) {
-        result *= 1i * k1_ *
-                  (incidentPhasor_ * std::exp(1i * k1_ * x) -
-                   reflectionPhasor_ * std::exp(-1i * k1_ * x));
-    } else {
-        result *= 1i * k2_ * transmissionPhasor_ * std::exp(1i * k2_ * x);
-    }
-    return result;
-}
-
-template <std::size_t dim>
-BoundaryConditionType PlaneWaveReflectionProblem<dim>::getBoundaryConditionType(
-    const Base::Face &face) const {
-    auto normal = face.getNormalVector(
-        face.getReferenceGeometry()->getCenter().castDimension<dim - 1>());
-    normal /= normal.l2Norm();
-
-    if (std::abs(normal[0]) > 0.9) {
-        // Normal in x direction
-        return BoundaryConditionType::SILVER_MULLER;
-    } else if (std::abs(normal[1]) > 0.9) {
-        // Normal in y-direction, E-field parallel to the normal
-        // use perfect electric conductor
-        return BoundaryConditionType::DIRICHLET;
-    } else if (dim > 2 && std::abs(normal[2]) > 0.9) {
-        // Normal in z direction, field parallel to the face
-        // use perfect magnetic conductor
-        return BoundaryConditionType::NEUMANN;
-    } else {
-        logger.assert_always(
-            false, "Problem not set up for a face in normal direction %",
-            normal);
-        return BoundaryConditionType::DIRICHLET;
-    }
-}
-template <std::size_t dim>
-LinearAlgebra::SmallVectorC<dim>
-    PlaneWaveReflectionProblem<dim>::boundaryCondition(
-        Base::PhysicalFace<dim> &face) const {
-    return ExactHarmonicProblem<dim>::boundaryCondition(face);
-}
-
-template class PlaneWaveReflectionProblem<2>;
-template class PlaneWaveReflectionProblem<3>;
 }  // namespace DGMax
